Move phan/phan to composer require and upgrade it
The 2.2.13 version (still not released) notably brings in an improvement
with call_user_func_array [1]. We need that to avoid tons of false
positives for hooks taking a pass-by-ref, and to be able to enable
redundant_condition_detection in core.

Bonus: upgrade PHPUnit; after all, it's not used by this repo (yet).

[1] - https://github.com/phan/phan/issues/3186

Release notes:
https://github.com/phan/phan/releases/tag/2.2.12
https://github.com/phan/phan/releases/tag/2.2.13

Bug: T220589
Change-Id: I7d488bc3eebddf7e8562355e76071ed20b3d827a