Add `isset()` checks to RedundantExistenceChecksPlugin
isset() is very similar to empty(), in that it should never be used on
expressions that are known to be set, because it provides no advantages,
and instead makes the code less robust by potentially hiding mistyped
variable names.

The only difference with empty() is that isset() can only be used on
variable- and property-like expressions, so we don't need to test every
node type.

Change-Id: Idad810322b4ec959a2d4b71c88ab496a0e2ef447