Fix handling of dynamic and doc-only properties
If phan already knows that a property is dynamic or from a PHPDoc, we
should never emit an issue. We don't even need to check the property's
class to do that.

Also fix an issue where isset() on non-dynamic properties would never be
flagged as incorrect when the class allowed dynamic properties (even if
the property in question wasn't dynamic in the first place).

Bug: T382208
Change-Id: Id0448bd40ca7446a7ae7e1182bd2f0dfd78a77de