Add FirstClassCallableRecommendPlugin
Recommends replacing callable strings and arrays with first-class
callable syntax. Also includes a fixer for `phan --automatic-fix`!

  call( 'func' );                  →  call( func( ... ) );
  call( [ 'Cls', 'func' ] );       →  call( Cls::func( ... ) );
  call( [ Cls::class, 'func' ] );  →  call( Cls::func( ... ) );
  call( [ $x::class, 'func' ] );   →  call( $x::func( ... ) );
  call( [ __CLASS__, 'func' ] );   →  call( self::func( ... ) );
  call( [ $obj, 'func' ] );        →  call( $obj->func( ... ) );
  call( [ $str, 'func' ] );        →  call( $str::func( ... ) );

Issues are emitted when a literal string/array that is a valid
callable is either passed as a method parameter or assigned to a
property which is type-hinted as a callable.

The MediaWikiUseFirstClassCallable issue type is enabled in our
config, but MediaWikiUseFirstClassCallableInternalFunc is suppressed,
since using the `strlen( ... )` syntax for built-in functions feels
quite verbose.

Bug: T397107
Change-Id: I54abd989d457b316805a21b7ee9b9bd152bfd8c0