Do not require `@throws` tags for the `Error` hierarchy
Quoting the PHP manual (https://www.php.net/manual/en/class.error.php):

> Error is the base class for all internal PHP errors.

This means that userland code should never throw it in the first place.
However, some places still throw Error (occasionally a subclass)
directly.

It should be safe to assume that this is only done for unrecoverable
errors; else the code should use a proper exception, but that's out of
scope for this change. So, do not require `@throws` tag for it.

Bug: T338426
Change-Id: Id2d4ee49ebc4a2182a5ccc26d82d72221a4ea359